file structure created with architecture notes, landingHeader component inserted and basic router inserted
