+++ src/app/App.jsx
@@ -1,0 +1,27 @@
+import React from 'react';
+import {
+  BrowserRouter as Router,
+  Route,
+  Switch
+} from 'react-router-dom';
+
+
+
+export default function App() {
+  return (
+    <>
+      <Router>
+
+        <Switch>
+
+          <Route 
+            path="/"
+            exact
+            component= />
+
+        </Switch>
+
+      </Router>
+    </>
+  );
+}
