+++ src/components/LandingHeader.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+
+function LandingHeader() {
+  return (
+    <>
+      <header>
+        <img src={image} alt={name} />
+        
+      </header>    
+    </>
+  );
+}
+
+export default LandingHeader;
