--- src/components/app/App.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export default function App() {
-  return <h1>Hello World</h1>;
-}
