--- src/components/app/App.test.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import { render, cleanup } from '@testing-library/react';
-import App from './App';
-
-describe('App component', () => {
-  afterEach(() => cleanup());
-  it('renders App', () => {
-    const { asFragment } = render(<App />);
-    expect(asFragment()).toMatchSnapshot();
-  });
-});
