--- src/setupTests.js
@@ -1,1 +1,0 @@
-global.fetch = require('node-fetch');
