+++ src/utils/setupTests.js
@@ -1,0 +1,1 @@
+global.fetch = require('node-fetch');
